+++ mvlearn/__init__.py
@@ -5,6 +5,8 @@ import mvlearn.semi_supervised
 import mvlearn.datasets
 import mvlearn.utils
 import mvlearn.decomposition
+import mvlearn.merge
 import mvlearn.plotting  # noqa
+
 
 __version__ = "0.3.0"